moved spinner & logger to new files
